Another for-loop index type change

+++ 2/1/main.c
@@ -25,7 +25,7 @@ typedef enum find_mode
 int count_commas(char* line)
 {
     int commas = 0;
-    for (int i = 0; i < strlen(line); i++)
+    for (size_t i = 0; i < strlen(line); i++)
     {
         if (line[i] == ',')
         {
